Update PoisonToads.sol
added eirc155 check -- working on VM!!  discountCost reflected the change based on whether a wallet had a utility toad token in it or not.  Next steps are to get the three levels of discount in.  then edit mint function to use the discounted price.  then deploy and test on mumbai.